Refactor and abstract search capabilities
