+++ app/controllers/api/v1/users_controller.rb
@@ -1,4 +1,7 @@
 class Api::V1::UsersController < Api::V1::ApiController
+  include Search
+  include UserProfileSearch
+
   before_action :set_user, only: [:show, :update, :destroy]
 
   PERMITTED_ATTRIBUTES = [
@@ -30,7 +33,8 @@ class Api::V1::UsersController < Api::V1::ApiController
     users = User.eager_load(ASSOCIATIONS).all
     users = filter(users) if search_params.to_h.any?
     users = profile_filter(users) if profile_search_params.to_h.any?
-    users = fuzzy_filter(users) if fuzzy_search_params.to_h.any?
+    #users = fuzzy_filter(users) if fuzzy_search_params.to_h.any?
+    users = profile_fuzzy_filter(users) if profile_fuzzy_search_params.to_h.any?
     render_paginated(users)
   end
 
@@ -64,46 +68,19 @@ private
     params.require(:user).permit(PERMITTED_ATTRIBUTES)
   end
 
-  #
-  # SEARCH
-  #
-
   def search_params
     params.permit([:email, :username, :confirmed, :visible, :role, :access_level, :customer_uuid])
   end
+
+  #def fuzzy_search_params
+  #  params.permit(fuzzy: [:email, :username])
+  #end
 
   def profile_search_params
     params.permit([:first_name, :last_name])
   end
 
-  def filter(resources)
-    resources.where(search_params.to_h)
-  end
-
-  def profile_filter(resources)
-    profile_search_params.to_h.each do |k,v|
-      resources = resources.where("user_profiles.#{k} = ?", v)
-    end
-    return resources
-  end
-
-  #
-  # FUZZY SEARCH
-  #
-
-  def fuzzy_search_params
-    params.permit(fuzzy: [:name])
-  end
-
-  def fuzzy_filter(resources)
-    fuzzy_search_params.to_h["fuzzy"].each do |k,v|
-      if k == "name"
-        resources = resources.where("user_profiles.first_name ILIKE :name OR user_profiles.last_name ILIKE :name", name: "%#{v}%")
-      else
-        resources = resources.where("#{k} ILIKE ?", "%#{v}%")
-      end
-    end
-
-    return resources
+  def profile_fuzzy_search_params
+    params.permit(fuzzy: [:name, :first_name, :last_name])
   end
 end
